requirement.txt
google.generativeai==0.4.1
gradio
gdown
transformers==4.30.2
streamlit